fix(atom/tag): fit tests
