+++ components/atom/tag/src/constants.js
@@ -1,0 +1,30 @@
+export const ACTIONABLE_ONLY_PROPS = [
+  'href',
+  'iconPlacement',
+  'target',
+  'actionable',
+  'linkFactory',
+  'rel'
+]
+
+export const STANDARD_ONLY_PROPS = ['closeIcon', 'onClose']
+
+export const SIZES = {
+  LARGE: 'large',
+  MEDIUM: 'medium',
+  SMALL: 'small'
+}
+
+export const LINK_TYPES = {
+  NOFOLLOW: 'nofollow',
+  NOOPENER: 'noopener',
+  NOREFERRER: 'noreferrer',
+  PREV: 'prev',
+  NEXT: 'next',
+  TAG: 'tag'
+}
+
+export const DESIGNS = {
+  SOLID: 'solid',
+  OUTLINE: 'outline'
+}
